+++ src/ReconNess.UnitTests/AgentServiceTests.cs
@@ -1,0 +1,38 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace ReconNess.UnitTests
+{
+    [TestClass]
+    public class AgentServiceTests
+    {
+        [TestMethod]
+        public void TestRunAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+
+        [TestMethod]
+        public void TestStopAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+
+        [TestMethod]
+        public void TestDebugAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+    }
+}