+++ src/ReconNess.UnitTests/NotesServiceTests.cs
@@ -1,0 +1,28 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace ReconNess.UnitTests
+{
+    [TestClass]
+    public class NotesServiceTests
+    {
+        [TestMethod]
+        public void TestSaveTargetNotesAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+
+        [TestMethod]
+        public void TestSaveSubdomainNotesAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+    }
+}