+++ src/ReconNess.UnitTests/ReferenceServiceTests.cs
@@ -1,0 +1,18 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace ReconNess.UnitTests
+{
+    [TestClass]
+    public class ReferenceServiceTests
+    {
+        [TestMethod]
+        public void TestGetAllCategoriesAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+    }
+}