+++ src/ReconNess.UnitTests/SubdomainServiceTests.cs
@@ -1,0 +1,28 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace ReconNess.UnitTests
+{
+    [TestClass]
+    public class SubdomainServiceTests
+    {
+        [TestMethod]
+        public void TestUpdateSubdomainAsyncMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+
+        [TestMethod]
+        public void TestDeleteSubdomainsMethod()
+        {
+            // Arrange
+
+            // Act
+
+            // Assert
+        }
+    }
+}