+++ src/ReconNess/Services/AgentService.cs
@@ -95,6 +95,7 @@ namespace ReconNess.Services
                     var needToSkip = agent.SkipIfRanBefore && (!string.IsNullOrEmpty(sub.FromAgents) && sub.FromAgents.Contains(agent.Name));
                     if (needToBeAlive || needToSkip)
                     {
+                        await this.connectorService.SendAsync("Logs_" + channel, $"Skip subdomain: {sub.Name}");
                         continue;
                     }
 
@@ -175,7 +176,7 @@ namespace ReconNess.Services
                     await this.connectorService.SendAsync("Logs_" + channel, $"Output: {terminalLineOutput}");
                     await this.connectorService.SendAsync("Logs_" + channel, $"Result: {JsonConvert.SerializeObject(scriptOutput)}");
 
-                    await this.targetService.SaveScriptOutputAsync(target, subdomain, agent, scriptOutput, subdomain == null, cancellationToken);
+                    await this.targetService.SaveScriptOutputAsync(target, subdomain, agent, scriptOutput, cancellationToken);
 
                     await this.connectorService.SendAsync("Logs_" + channel, $"Output #: {lineCount} processed");
                     await this.connectorService.SendAsync("Logs_" + channel, "-----------------------------------------------------");